temperature/requirements.txt
paho-mqtt==2.1.0
typing-extensions==4.7.1
